+++ pkgs/native_assets_builder/test_data/add_asset_link/hook/build.dart
@@ -26,7 +26,7 @@ void main(List<String> arguments) async {
       input: input,
       output: output,
       logger: logger,
-      linkInPackage: 'add_asset_link',
+      routing: const [ToLinkHook('add_asset_link')],
     );
   });
 }