+++ pkgs/native_assets_builder/test_data/complex_link/hook/build.dart
@@ -27,7 +27,10 @@ void main(List<String> args) async {
 
       output.assets.data.add(
         DataAsset(package: packageName, name: name, file: dataAsset.uri),
-        linkInPackage: input.config.linkingEnabled ? packageName : null,
+        routing:
+            input.config.linkingEnabled
+                ? ToLinkHook(packageName)
+                : const ToAppBundle(),
       );
       // TODO(https://github.com/dart-lang/native/issues/1208): Report
       // dependency on asset.