+++ pkgs/native_assets_builder/test_data/fail_on_os_sdk_version_link/hook/build.dart
@@ -12,8 +12,10 @@ void main(List<String> arguments) async {
         file: input.packageRoot.resolve('assets/data.json'),
         package: input.packageName,
       ),
-      linkInPackage:
-          input.config.linkingEnabled ? 'fail_on_os_sdk_version_linker' : null,
+      routing:
+          input.config.linkingEnabled
+              ? const ToLinkHook('fail_on_os_sdk_version_linker')
+              : const ToAppBundle(),
     );
   });
 }