+++ pkgs/native_assets_builder/test_data/native_add_duplicate/hook/build.dart
@@ -32,7 +32,10 @@ void main(List<String> arguments) async {
     output.assets.code.add(
       tempBuildOutput.assets.code.single,
       // Send dylib to linking if linking is enabled.
-      linkInPackage: input.config.linkingEnabled ? packageName : null,
+      routing:
+          input.config.linkingEnabled
+              ? ToLinkHook(packageName)
+              : const ToAppBundle(),
     );
     output.addDependencies(tempBuildOutput.dependencies);
   });