+++ pkgs/native_assets_cli/lib/src/code_assets/config.dart
@@ -151,16 +151,17 @@ extension CodeAssetBuildOutputBuilder on EncodedAssetBuildOutputBuilder {
 extension type CodeAssetBuildOutputBuilderAdd._(
   EncodedAssetBuildOutputBuilder _output
 ) {
-  /// Adds the given [asset] to the hook output (or send to [linkInPackage]
-  /// for linking if provided).
-  void add(CodeAsset asset, {String? linkInPackage}) =>
-      _output.addEncodedAsset(asset.encode(), linkInPackage: linkInPackage);
+  /// Adds the given [asset] to the hook output with [routing].
+  void add(CodeAsset asset, {AssetRouting routing = const ToAppBundle()}) =>
+      _output.addEncodedAsset(asset.encode(), routing: routing);
 
-  /// Adds the given [assets] to the hook output (or send to [linkInPackage]
-  /// for linking if provided).
-  void addAll(Iterable<CodeAsset> assets, {String? linkInPackage}) {
+  /// Adds the given [assets] to the hook output with [routing].
+  void addAll(
+    Iterable<CodeAsset> assets, {
+    AssetRouting routing = const ToAppBundle(),
+  }) {
     for (final asset in assets) {
-      add(asset, linkInPackage: linkInPackage);
+      add(asset, routing: routing);
     }
   }
 }