+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,6 +1,7 @@
-## 0.13.1-wip
+## 0.14.0-wip
 
-- Nothing yet.
+- Bump `package:native_assets_cli` to 0.14.0.
+- Route assets from build hook to build hook with `ToBuild` `Routing`.
 
 ## 0.13.0